Feat: Implement monitoring new line over in report message
